+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [BREAKING] Replaced `Action.config.default_log_level` and `default_autolog_level` with simpler `log_level`
 * [BREAKING] `autolog_level` method overrides with e.g. `auto_log :warn` or `auto_log false`
 * [BREAKING] Direct access to exposed fields in callables no longer works -- `foo` becomes `result.foo`
+* [BREAKING] Removed `success?` check on Action::Result (use `ok?` instead)
 
 ## 0.1.0-alpha.2.6
 * Inline interactor code (no more dependency on unpublished forked branch to support inheritance)